32-bit number of exceptions (hasindu-inspired)

+++ press/press.h
@@ -436,7 +436,7 @@ void turbopfor_depress_16(uint8_t *in, uint64_t nin, int16_t *out,
 
 /* variable byte 1 except 2 */
 
-#define VB1E2_MAX_EXCEPTIONS (UINT16_MAX)
+#define VB1E2_MAX_EXCEPTIONS (UINT32_MAX)
 uint64_t vb1e2_bound(uint32_t nin);
 void vb1e2_press(const uint16_t *in, uint32_t nin, uint8_t *out,
 		 uint64_t *nout);
